+++ TradingBot/Dto/StrategyParametersResponse.cs
@@ -11,5 +11,5 @@ public sealed class StrategyParametersResponse
     [Required]
     public required int MinDaysIncreasing { get; init; }
     [Required]
-    public required decimal TopGrowingSymbolsBuyRatio { get; init; }
+    public required double TopGrowingSymbolsBuyRatio { get; init; }
 }
